+++ lib/managers/config_manager.dart
@@ -5,9 +5,9 @@ import 'package:adapters_flutter/models/config/file_config_model.dart';
 import 'package:adapters_flutter/models/config/merged_config_model.dart';
 import 'package:adapters_flutter/services/config/env_config_service.dart';
 import 'package:adapters_flutter/services/config/file_config_service.dart';
+import 'package:adapters_flutter/services/validation_service.dart';
 import 'package:path/path.dart' as path;
 import 'package:synchronized/synchronized.dart';
-import 'package:uuid/uuid.dart';
 
 MergedConfigModel? _config;
 final _lock = Lock();
@@ -20,14 +20,14 @@ Future<MergedConfigModel> getConfigAsync() async {
       final envConfig = await getConfigFromEnvAsync();
 
       _config = _mergeConfigs(fileConfig, envConfig);
-      _validateConfig(_config);
+      validateConfig(_config);
     }
   });
 
   return _config as MergedConfigModel;
 }
 
-Future<void> updateTestRunIdAsync(String testRunId) async {
+Future<void> updateTestRunIdAsync(final String testRunId) async {
   (await getConfigAsync()).testRunId = testRunId;
 }
 
@@ -94,44 +94,7 @@ MergedConfigModel _mergeConfigs(
   return config;
 }
 
-void _validateConfig(final MergedConfigModel? config) async {
-  if (config == null) {
-    throw ArgumentError('Config is null');
-  }
-
-  if (config.adapterMode == 0 || config.adapterMode == 1) {
-    if (config.testRunId == null ||
-        !Uuid.isValidUUID(fromString: config.testRunId!)) {
-      throw FormatException('TestRunID is invalid');
-    }
-  } else if (config.adapterMode == 2) {
-    if (config.testRunId != null && config.testRunId!.isNotEmpty) {
-      throw FormatException('TestRunID should be absent in adapter mode 2');
-    }
-  } else {
-    throw FormatException('Invalid adapter mode: ${config.adapterMode}');
-  }
-
-  if (config.projectId == null ||
-      !Uuid.isValidUUID(fromString: config.projectId!)) {
-    throw FormatException('ProjectId is invalid');
-  }
-
-  if (config.configurationId == null ||
-      !Uuid.isValidUUID(fromString: config.configurationId!)) {
-    throw FormatException('ConfigurationId is invalid');
-  }
-
-  if (config.privateToken == null || config.privateToken!.isEmpty) {
-    throw FormatException('PrivateToken is invalid');
-  }
-
-  if (config.url == null || !Uri.parse(config.url!).isAbsolute) {
-    throw FormatException('Url is invalid');
-  }
-}
-
-MergedConfigModel _updateUrl(MergedConfigModel config) {
+MergedConfigModel _updateUrl(final MergedConfigModel config) {
   if (config.url?.endsWith('/') ?? false) {
     config.url = config.url!.substring(0, config.url!.length - 1);
   }