+++ lib/models/api/attachment_api_model.dart
@@ -5,22 +5,20 @@ final class AttachmentPutModel {
 
   const AttachmentPutModel(this.id);
 
-  Map<String, dynamic> toJson() => {
-        'id': id,
-      };
+  Map<String, dynamic> toJson() => {'id': id};
 }
 
-final class AttachmentPutModelAutoTestStepResultsModel {
-  final Outcome? outcome;
-  final String? title;
-  final String? description;
-  final String? info;
-  final DateTime? startedOn;
-  final DateTime? completedOn;
-  final int? duration;
-  final List<AttachmentPutModelAutoTestStepResultsModel>? stepResults;
-  final List<AttachmentPutModel>? attachments;
-  final Map<String, String>? parameters;
+final class AutoTestStepResultsModel {
+  Outcome? outcome;
+  String? title;
+  String? description;
+  String? info;
+  DateTime? startedOn;
+  DateTime? completedOn;
+  int? duration;
+  List<AutoTestStepResultsModel> stepResults = [];
+  List<AttachmentPutModel> attachments = [];
+  Map<String, String>? parameters = {};
 
   Map<String, dynamic> toJson() => {
         'outcome': outcome?.name,
@@ -32,20 +30,8 @@ final class AttachmentPutModelAutoTestStepResultsModel {
         'duration': duration,
         'stepResults': stepResults,
         'attachments': attachments,
-        'parameters': parameters,
+        'parameters': parameters
       };
-
-  const AttachmentPutModelAutoTestStepResultsModel(
-      this.outcome,
-      this.title,
-      this.description,
-      this.info,
-      this.startedOn,
-      this.completedOn,
-      this.duration,
-      this.stepResults,
-      this.attachments,
-      this.parameters);
 }
 
 final class AttachmentResponseModel {