+++ lib/models/api/label_api_model.dart
@@ -10,7 +10,5 @@ final class LabelPostModel {
 
   const LabelPostModel(this.name);
 
-  Map<String, dynamic> toJson() => {
-        'name': name,
-      };
+  Map<String, dynamic> toJson() => {'name': name};
 }