+++ lib/models/test_result_model.dart
@@ -5,9 +5,9 @@ import 'package:adapters_flutter/models/api/link_api_model.dart';
 final class TestResultModel {
   String? namespace;
   String? classname;
-  List<AttachmentPutModelAutoTestStepResultsModel> steps = [];
-  List<AttachmentPutModelAutoTestStepResultsModel> setup = [];
-  List<AttachmentPutModelAutoTestStepResultsModel> teardown = [];
+  List<AutoTestStepResultsModel> steps = [];
+  List<AutoTestStepResultsModel> setup = [];
+  List<AutoTestStepResultsModel> teardown = [];
   String? externalId;
   String? name;
   String? title;
@@ -24,6 +24,6 @@ final class TestResultModel {
   DateTime? completedOn;
   int? duration;
   List<AttachmentPutModel> attachments = [];
-  Map<String, String>? parameters;
-  Map<String, String>? properties;
+  Map<String, String> parameters = {};
+  Map<String, String> properties = {};
 }