+++ lib/services/api/autotest_api_service.dart
@@ -5,7 +5,8 @@ import 'package:adapters_flutter/converters/test_result_converter.dart';
 import 'package:adapters_flutter/models/api/autotest_api_model.dart';
 import 'package:adapters_flutter/models/api/workitem_api_model.dart';
 import 'package:adapters_flutter/models/config/merged_config_model.dart';
-import 'package:adapters_flutter/models/test_result.dart';
+import 'package:adapters_flutter/models/exception_model.dart';
+import 'package:adapters_flutter/models/test_result_model.dart';
 import 'package:http/http.dart';
 import 'package:logger/logger.dart';
 
@@ -26,8 +27,7 @@ Future<AutotestFullModel?> createAutotestAsync(
     final request =
         Request('POST', Uri.parse('${config.url}/api/v2/autoTests'));
 
-    final requestBody =
-        await toCreateAutotestRequestModelAsync(config, testResult);
+    final requestBody = toCreateAutotestRequestModel(config, testResult);
     requestBody.shouldCreateWorkItem =
         config.automaticCreationTestCases ?? false;
     request.body = json.encode(requestBody);
@@ -37,7 +37,7 @@ Future<AutotestFullModel?> createAutotestAsync(
     final response = await Response.fromStream(streamedResponse);
 
     if (response.statusCode < 200 || response.statusCode > 299) {
-      throw HttpException(
+      throw TmsApiException(
           'Status code: ${response.statusCode}, Reason: ${response.reasonPhrase}');
     }
 
@@ -85,7 +85,7 @@ Future<AutotestFullModel?> getAutotestByExternalIdAsync(
     final response = await Response.fromStream(streamedResponse);
 
     if (response.statusCode < 200 || response.statusCode > 299) {
-      throw HttpException(
+      throw TmsApiException(
           'Status code: ${response.statusCode}, Reason: ${response.reasonPhrase}');
     }
 
@@ -118,7 +118,7 @@ Future<bool> tryLinkAutoTestToWorkItemAsync(final MergedConfigModel config,
       final response = await request.send();
 
       if (response.statusCode < 200 || response.statusCode > 299) {
-        throw HttpException(
+        throw TmsApiException(
             'Status code: ${response.statusCode}, Reason: ${response.reasonPhrase}');
       }
     } catch (exception, stacktrace) {
@@ -142,15 +142,14 @@ Future<void> updateAutotestAsync(
     };
 
     final request = Request('PUT', Uri.parse('${config.url}/api/v2/autoTests'));
-    final requestBody =
-        await toUpdateAutotestRequestModelAsync(config, testResult);
+    final requestBody = toUpdateAutotestRequestModel(config, testResult);
     request.body = json.encode(requestBody);
     request.headers.addAll(headers);
 
     final response = await request.send();
 
     if (response.statusCode < 200 || response.statusCode > 299) {
-      throw HttpException(
+      throw TmsApiException(
           'Status code: ${response.statusCode}, Reason: ${response.reasonPhrase}');
     }
   } catch (exception, stacktrace) {