+++ test/adapters_flutter_test.dart
@@ -1,29 +1,54 @@
 import 'package:adapters_flutter/adapters_flutter.dart';
-import 'package:adapters_flutter/managers/config_manager.dart';
+import 'package:adapters_flutter/enums/link_type_enum.dart';
+import 'package:adapters_flutter/models/api/link_api_model.dart';
+import 'package:flutter_test/flutter_test.dart';
 
 void main() async {
-  await testAsync('calculate', externalId: 'lalalala', workItemsIds: ['45812'],
-      () async {
-    //await addAttachmentAsync('avatar.png');
-    await stepAsync('example step title', () async {
-      await getConfigAsync();
-      await getConfigAsync();
-    });
-
-    await stepAsync('failed step', () async {
-      throw Exception('example exception');
-    });
+  setUpAll(() {
+    print('setUpAll');
   });
 
-  await testAsync('calculate2',
-      externalId: 'fdgdfgdfgdf', workItemsIds: ['45812'], () async {
-    await stepAsync('example step title', () async {
-      await getConfigAsync();
-      await getConfigAsync();
+  setUp(() {
+    print('setUp');
+  });
+
+  await tmsTest('example test',
+      externalId: 'example_test',
+      title: 'example_title',
+      tags: [
+        'example_tag'
+      ],
+      links: [
+        const Link('link_title', 'https://www.example.org/', 'link_description',
+            LinkType.issue)
+      ],
+      workItemsIds: [
+        '45835'
+      ], () async {
+    await step('success step', () => expect(0, 0));
+
+    await step('success step with attachment',
+        () async => await addAttachment('avatar.png'));
+
+    await step('success step with message',
+        () async => await addMessage('example message'));
+
+    await step('success step with link',
+        () async => await addLink('https://www.example.org/'));
+
+    await step('success step with body', () {
+      const actual = 0;
+      expect(actual, 0);
     });
 
-    await stepAsync('failed step', () async {
-      throw Exception('example exception');
-    });
+    await step('failed step', () => throw Exception('example exception'));
+  });
+
+  tearDown(() {
+    print('tearDown');
+  });
+
+  tearDownAll(() {
+    print('tearDownAll');
   });
 }