+++ CHANGELOG.md
@@ -3,6 +3,27 @@
 All notable changes to this project will be documented in this file.  
 The format is based on Keep a Changelog and this project adheres to Semantic Versioning.
 
+## [1.0.0] - 2024-06-26
+
+### Changed
+
+- public functions names.
+
+### Fixed
+
+- readme example.
+
+## [0.1.0] - 2024-06-25
+
+### Added
+
+- child steps support
+
+### Fixed
+
+- http LinkType request body encode.
+- labels argument name.
+
 ## [0.1.2] - 2024-06-24
 
 ### Fixed